% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{RcppKyTea-package}
\alias{RcppKyTea}
\alias{RcppKyTea-package}
\title{RcppKyTea: 'Rcpp' Interface to 'KyTea'}
\description{
'RcppKyTea' just tokenizes a character vector into tokens (a list of character vector). That's all. This package does not provide interface to APIs for training models.
}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

Other contributors:
\itemize{
  \item The LIBLINEAR Project [copyright holder]
  \item KyTea Development Team [copyright holder]
}

}
\keyword{internal}
