% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize.R
\name{tokenize}
\alias{tokenize}
\title{Tokenize text using KyTea}
\usage{
tokenize(text, model = Sys.getenv("KYTEA_MODEL"))
}
\arguments{
\item{text}{Character vector.}

\item{model}{Path to model file.}
}
\value{
List of character vectors.
}
\description{
Tokenize text using KyTea
}
